backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
stripe==12.5.0
PyJWT>=2.8.0
google-auth>=2.29.0
requests>=2.31.0
httpx>=0.27.0
